+++ libffi-rs/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [5.2.0] - 2026-08-15
 
 - Add fallible CIF, builder, and closure constructors.
 - Add support for getting the offsets of a struct via `low::get_struct_offsets` or `middle::Type::struct_offsets`.